+++ app/src/main/java/com/example/f1project/ui/profile/ConstructorProfileScreen.kt
@@ -98,7 +98,7 @@ fun ConstructorProfileScreen(
                 item { ConstructorInfoCard(profile = profile) }
                 item {
                     SeasonStatsCard(
-                        points = profile.seasonStats.points,
+                        points = profile.seasonStats.points.toInt(),
                         wins = profile.seasonStats.wins,
                         podiums = profile.seasonStats.podiums
                     )
